% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_lengths}
\alias{chr_lengths}
\title{Data on chromosome lengths.}
\format{
A numeric vector named by the chromosome number.
}
\source{
\url{https://plasmodb.org/plasmo/showApplication.do}
}
\usage{
chr_lengths
}
\description{
Lengths in base pairs of chromosomes Pf3D7_01_v3 to Pf3D7_14_v3 of the 3D7
Plasmodium falciparum reference genome listed on PlasmoDB (see url below).
}
\keyword{datasets}
