% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_marker_summaries.R
\name{compute_diversities}
\alias{compute_diversities}
\title{Compute marker diversities}
\usage{
compute_diversities(frequencies)
}
\arguments{
\item{frequencies}{ndata (marker count) by Kmax (max cardinality over m
markers) matrix if Kt < Kmax for any t in 1:m, then fs[t,1:Kt] in (0,1) &
fs[t,(Kt+1):Kmax] = 0 Example: if Kt = 2 < Kmax = 4 then fs[t,] might look
like [0.2, 0.7, 0, 0].}
}
\value{
Diversities, \eqn{h_t}, for \eqn{t = 1,\ldots,m}
  markers.
}
\description{
Given a matrix frequencies \code{compute_diversities()} returns the
diversities, \eqn{h_t}, for \eqn{t = 1,\ldots,m} markers. Each
calculated as described in Taylor et al. 2009
without accounting for sample bias:
\deqn{h_t^{'} = \dfrac{1}{h_t}}
}
\examples{
compute_diversities(fs = frequencies$Colombia)

}
\references{
Taylor, A.R., Jacob, P.E., Neafsey, D.E. and Buckee, C.O., 2019.
  Estimating relatedness between malaria parasites. Genetics, 212(4),
  pp.1337-1351.
}
