% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frequencies}
\alias{frequencies}
\title{Allele frequencies of the example GTseq panel.}
\format{
Each entry of the list is a matrix, \code{fs} say, with \eqn{m=126}
  rows and \eqn{Kmax=44} variables, where \eqn{m} is the marker count and
  \eqn{Kmax} is the maximum cardinality (per-marker allele count) observed
  over all \eqn{m} markers. If, for any \eqn{t = 1,...,m}, the maximum
  cardinality exceeds that of the \eqn{t}-th marker (i.e. if \eqn{Kmax >
  Kt}), then all \code{fs[t,1:Kt]} are in (0,1) and all
  \code{fs[t,(Kt+1):Kmax]} are zero. For example, for PF3D7_0103600 in Colombia, \eqn{Kt
  = 2} and \code{frequencies$Colombia["PF3D7_0103600",] = (0.687075, 0.312925, 0, ..., 0)}.
  \describe{
  \item{Allele.1}{Frequency (numeric) of the first allele}
  ...
  \item{Allele.44}{Frequency (numeric) of the \eqn{Kmax} allele}
}
}
\source{
see \url{https://github.com/artaylor85/paneljudge/blob/master/data_raw/Process_GTseq.R}
}
\usage{
frequencies
}
\description{
A data set of allele frequencies for four countries: Colombia, French Guiana, Mali and Sengal.
}
\keyword{datasets}
