% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Data on markers of the example GTseq panel.}
\format{
A data frame with 126 rows and 8 variables:
\describe{
  \item{Amplicon_name}{Name (character) of the microhaplotype marker ("Amplicon" because typed using an amplicon)}
  \item{Chr}{Chromosome (character) of the microhaplotype marker}
  \item{Start}{First base pair (integer) of the microhaplotype marker}
  \item{Stop}{Last base pair  (integer) of the microhaplotype marker}
  \item{length}{Length (integer) of the microhaplotype marker in base pairs}
  \item{pos}{Mid-point (numeric) of the microhaplotype marker}
  \item{chrom}{Chromosome (numeric) of the microhaplotype marker}
  \item{distance}{Inter mid-point distance (numeric) between the microhaplotype marker and its subsequent neighbour}
}
}
\source{
see \url{https://github.com/artaylor85/paneljudge/blob/master/data_raw/Process_GTseq.R}
}
\usage{
markers
}
\description{
A data set of marker attributes for markers pertaining to the example GTseq panel.
}
\keyword{datasets}
