% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data on a pair of haploid genotypes}
\usage{
simulate_data(fs, gendist, k, r, epsilon = 0.001, rho = 7.4 * 10^(-7))
}
\arguments{
\item{fs}{frequencies. ndata (marker count) by Kmax (max cardinality over m
markers) matrix if Kt < Kmax for any t in 1:m, then fs[t,1:Kt] in (0,1) &
fs[t,(Kt+1):Kmax] = 0 Example: if Kt = 2 < Kmax = 4 then fs[t,] might look
like [0.2, 0.7, 0, 0].}

\item{gendist}{distances vector where gendist[t] contains the distance
between position t and t+1 or equivalently, gendist[t-1] contains the
distance between position t-1 and t, for p > 1. s.t. only m-1 first entries
of gendist are being used.}

\item{k}{data-generating switch rate parameter.}

\item{r}{data-generating relatedness parameter.}

\item{epsilon}{genotyping error (probability of miscall).}

\item{rho}{recombination rate in probability of break point per base pair.}
}
\value{
Ys simulated dated.
}
\description{
For a vector of input frequencies and distances, \code{simulate_data()}
returns simulated data on a pair of haploid genotypes.
}
\examples{
simulate_data(fs = frequencies$Colombia, gendist = markers$dt, k = 10, r = 0.5)

}
\references{
Taylor, A.R., Jacob, P.E., Neafsey, D.E. and Buckee, C.O., 2019.
  Estimating relatedness between malaria parasites. Genetics, 212(4),
  pp.1337-1351.
}
