% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_r_and_k_CIs.R
\name{compute_r_and_k_CIs}
\alias{compute_r_and_k_CIs}
\title{Compute confidence intervals for relatedness and switch rate parameters}
\usage{
compute_r_and_k_CIs(
  frequencies,
  distances,
  khat,
  rhat,
  confidence = 95,
  nboot = 100,
  do_not_use_core_count = 2,
  ...
)
}
\arguments{
\item{frequencies}{ndata (marker count) by Kmax (max cardinality over m
markers) matrix if Kt < Kmax for any t in 1:m, then fs[t,1:Kt] in (0,1) &
fs[t,(Kt+1):Kmax] = 0 Example: if Kt = 2 < Kmax = 4 then fs[t,] might look
like [0.2, 0.7, 0, 0].}

\item{distances}{vector where distances[t] contains the distance between
position t and t+1 or equivalently, gendist[t-1] contains the distance
between position t-1 and t, for p > 1. s.t. only m-1 first entries of
gendist are being used.}

\item{confidence}{confidence level (percentage) of confidence interval
(default 95\%)}

\item{nboot}{number of parametric bootstrap estimates from which to compute
the CI. Larger values provide a better approximation but CIs will take
longer to compute.}

\item{do_not_use_core_count}{number of cores to keep free on computer
(defaults to two)}

\item{epsilon}{genotyping error (probability of miscall).}

\item{rho}{recombination rate in probability of break point per base pair.}

\item{kinit}{switch rate parameter value used to initialise loglikelihood
optimization.}

\item{rinit}{relatedness parameter value used to initialise loglikelihood
optimization.}
}
\value{
Confidence interval bounds around input switch rate parameter, k, and
  relatedness parameter, r.
}
\description{
For specified confidence (default 95\%), given a matrix of input frequencies
and a vector of marker distances, a switch rate parameter estimate, khat, and
a relatedness parameter estimate, rhat. \code{compute_r_and_k_CIs()} returns
the confidence intervals around the parameter estimates. Confidence intervals
are approximate. They are approximated using the parametric bootstrap draws
under the HMM described in Taylor et al. 2009 The parametric bootstrap draws
are generated in parallel using a specifying number of cores. The quality and
compute time of the approximation scales with nboot.
}
\examples{
# First, stimulate some data
simulated_Ys <- simulate_Ys(fs = frequencies$Colombia, ds = markers$dt, k = 5, r = 0.25)

# Second, estimate the switch rate parameter, k, and relatedness parameter, r
krhat <- estimate_r_and_k(fs = frequencies$Colombia, ds = markers$dt, Ys = simulated_Ys)

# Third, compute confidence intervals (CIs)
compute_r_and_k_CIs(frequencies$Colombia, markers$dt, khat = krhat['khat'], rhat = krhat['rhat'])

}
\references{
Taylor, A.R., Jacob, P.E., Neafsey, D.E. and Buckee, C.O., 2019.
  Estimating relatedness between malaria parasites. Genetics, 212(4),
  pp.1337-1351.
}
